projects/sparse-rationale-constrained-attention-hate-speech/requirements.txt
torch>=2.1.0
transformers>=4.46.0
datasets>=2.18.0
hydra-core>=1.3.0
omegaconf>=2.3.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
